updated insertnews.py
corrected issue with timestamp. Some feeds will show duplicate news items when you fetch new items in RSSGuard, based on your timezone, as QuiteRSS didn't save the published date with timezone information. (It seems that RSSGuard does not unify news in Messages by their custom_id?)

changed how read/unread/deleted/pdeleted feeds are handled:

everything is marked as is_deleted = 0, so nothing goes to the recycle bin (this will add older news from QuiteRss as news with no author or other metadata, as that was regularly cleaned in QuiteRss)
everything that was deleted = 1 is marked as is_pdeleted =1 (so they are permanently deleted and nowhere to be seen)
everything that was deleted = 2 is marked as is_read = 1 (so older news from QuiteRss will be added as already read in your regular feeds)

more can be done to the first parts of the script to avoid doing by hand the work mentioned in https://github.com/martinrotter/rssguard/issues/1707#issuecomment-3315142577 and https://github.com/martinrotter/rssguard/issues/1707#issuecomment-3316128498, but I'll leave it for later. I'm sattisfied with the current outcome.